Fix issue #426: Split implementation_tests.rs into focused test modules
Automated fix by auto-issue-worker
