+++ docs/pr-summary-426.md
@@ -1,0 +1,92 @@
+## Summary
+
+Splits `implementation_tests.rs` (5,430 lines — the largest file in the codebase) into 10 focused test modules as part of Issue #426.
+
+### The Problem
+
+The monolithic `implementation_tests.rs` file had grown to 5,430 lines, making it difficult to:
+- Navigate and find specific tests
+- Understand test coverage by concern
+- Maintain and extend test suites
+- Run focused subsets of tests
+
+### Solution
+
+Reorganized tests into a modular directory structure at `src/analysis/implementation_tests/`:
+
+| Module | Lines | Description |
+|--------|-------|-------------|
+| `mod.rs` | 87 | Common imports and module declarations |
+| `cache_tests.rs` | 50 | Record cache contention handling |
+| `gpu_batch_tests.rs` | 100 | GPU batch evaluation edge cases |
+| `sample_matching_tests.rs` | 165 | Sample matching and filtering |
+| `optimal_weight_tests.rs` | 278 | Optimal outgoing weight calculation |
+| `relu_evaluation_tests.rs` | 322 | ReLU candidate evaluation and splitting |
+| `bias_calculation_tests.rs` | 332 | Bias calculation for various activation functions |
+| `improvement_model_tests.rs` | 385 | Linear and HARD_TANH improvement models |
+| `diagnostics_tests.rs` | 419 | Diagnostics and rejection tracking |
+| `prediction_accuracy_tests.rs` | 642 | Prediction accuracy vs manual simulation |
+| `synapse_analysis_tests.rs` | 1,284 | Synapse and neuron analysis integration |
+| **Total** | **4,064** | Split across 11 files (all under 1,500 lines) |
+
+### Key Decisions
+
+1. **Module structure over single file**: Used a directory with `mod.rs` rather than multiple standalone files for better organization
+
+2. **Common imports module**: Created a `common` submodule with shared imports to reduce duplication across test files
+
+3. **`$crate` for macro hygiene**: Updated the `skip_if_no_gpu!` macro to use `$crate` instead of `crate` for proper macro hygiene
+
+4. **Visibility restrictions**: Used `pub(super)` and `pub(crate)` to limit test module visibility appropriately
+
+### Files Changed
+
+| File | Change |
+|------|--------|
+| `src/analysis/implementation.rs` | Updated `#[path]` attribute to point to `implementation_tests/mod.rs` |
+| `src/analysis/implementation_tests.rs` | **Deleted** (5,430 lines) |
+| `src/analysis/implementation_tests/mod.rs` | **New** - module root with common imports |
+| `src/analysis/implementation_tests/cache_tests.rs` | **New** - record cache tests |
+| `src/analysis/implementation_tests/gpu_batch_tests.rs` | **New** - GPU batch tests |
+| `src/analysis/implementation_tests/sample_matching_tests.rs` | **New** - sample matching tests |
+| `src/analysis/implementation_tests/bias_calculation_tests.rs` | **New** - bias calculation tests |
+| `src/analysis/implementation_tests/diagnostics_tests.rs` | **New** - diagnostics tests |
+| `src/analysis/implementation_tests/relu_evaluation_tests.rs` | **New** - ReLU evaluation tests |
+| `src/analysis/implementation_tests/improvement_model_tests.rs` | **New** - improvement model tests |
+| `src/analysis/implementation_tests/synapse_analysis_tests.rs` | **New** - integration tests |
+| `src/analysis/implementation_tests/optimal_weight_tests.rs` | **New** - weight calculation tests |
+| `src/analysis/implementation_tests/prediction_accuracy_tests.rs` | **New** - prediction accuracy tests |
+
+Also cleaned up obsolete standalone test files that were created during development:
+- `src/analysis/record_cache_tests.rs` (deleted)
+- `src/analysis/gpu_batch_tests.rs` (deleted)
+- `src/analysis/sample_matching_tests.rs` (deleted)
+- `src/analysis/bias_calculation_tests.rs` (deleted)
+
+## Evidence
+
+Unable to generate screenshot: This is a CLI-only library with no visual interface. The refactoring is validated through the test suite.
+
+## Test Plan
+
+All 444 tests pass with `--test-threads=1`:
+
+```
+cargo test --lib -- --test-threads=1
+test result: ok. 444 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out
+```
+
+Quality checks pass:
+```
+./quality.sh
+✅ All quality checks passed!
+```
+
+### Acceptance Criteria Verification
+
+- [x] No test changes — purely reorganisation
+- [x] All tests still pass with `--test-threads=1`
+- [x] Maintain test coverage (444 tests, same as before)
+- [x] No single test file > 1,500 lines (max is 1,284 lines)
+- [x] Tests grouped by discovery module/concern
+- [x] `./quality.sh` passes
