+++ src/analysis/implementation_tests/synapse_analysis_tests.rs
@@ -1,0 +1,1284 @@
+//! Integration tests for synapse and neuron analysis functions.
+//!
+//! Extracted from implementation_tests.rs as part of Issue #426.
+//! Tests cover:
+//! - analyze_synapses validation and diagnostics
+//! - analyze_neurons validation and diagnostics
+//! - analyze_all combined analysis
+//! - Duplicate focus target rejection
+//! - Eligible source reporting
+//! - Deadline/timeout behaviour
+//! - Positive/negative improvement acceptance
+
+use super::common::*;
+
+#[test]
+fn analyze_neurons_rejects_duplicate_focus_targets() {
+    let temp_dir = tempdir().expect("Failed to create temporary directory");
+    let parquet_path = temp_dir.path().join("records.parquet");
+    let parquet_file = parquet_path
+        .to_str()
+        .expect("Temporary path should be valid UTF-8")
+        .to_string();
+
+    let sample_count = (MIN_NEURON_SAMPLE_COUNT + 5) as u32;
+    let mut records = Vec::new();
+    for obs_index in 0..sample_count {
+        records.push(DiscoverRecord::new(
+            obs_index,
+            "hidden-source".to_string(),
+            Some(0.0),
+            1.0,
+            vec![0.0],
+        ));
+        records.push(DiscoverRecord::new(
+            obs_index,
+            "output-0".to_string(),
+            Some(0.0),
+            0.5,
+            vec![1.0],
+        ));
+    }
+
+    write_records_to_parquet(&parquet_file, &records).expect("Failed to write discovery records");
+
+    let creature = CreatureJson {
+        input: 0,
+        output: 1,
+        neurons: vec![
+            NeuronJson {
+                uuid: "hidden-source".to_string(),
+                neuron_type: "hidden".to_string(),
+                squash: "IDENTITY".to_string(),
+                bias: 0.0,
+            },
+            NeuronJson {
+                uuid: "output-0".to_string(),
+                neuron_type: "output".to_string(),
+                squash: "IDENTITY".to_string(),
+                bias: 0.0,
+            },
+        ],
+        synapses: Vec::new(),
+    };
+
+    let input = AnalyzeNeuronsInput {
+        parquet_file: parquet_file.clone(),
+        creature,
+        focus_neurons: vec!["output-0".to_string(), "output-0".to_string()],
+        max_candidates: None,
+        analysis_deadline_ms: None,
+        random_seed: None,
+    };
+
+    let err =
+        analyze_neurons(&input).expect_err("Neuron analysis should refuse duplicate focus neurons");
+    let message = format!("{err}");
+    assert!(
+        message.contains("duplicate focus neurons"),
+        "Expected duplicate focus error, got: {message}",
+    );
+}
+
+#[test]
+fn analyze_synapses_rejects_duplicate_focus_targets() {
+    let temp_dir = tempdir().expect("Failed to create temporary directory");
+    let parquet_path = temp_dir.path().join("records.parquet");
+    let parquet_file = parquet_path
+        .to_str()
+        .expect("Temporary path should be valid UTF-8")
+        .to_string();
+
+    let sample_count = 16;
+    let mut records = Vec::new();
+    for obs_index in 0..sample_count {
+        records.push(DiscoverRecord::new(
+            obs_index,
+            "input-0".to_string(),
+            Some(0.0),
+            0.25,
+            vec![0.1],
+        ));
+        records.push(DiscoverRecord::new(
+            obs_index,
+            "output-0".to_string(),
+            Some(0.0),
+            0.5,
+            vec![-0.05],
+        ));
+    }
+
+    write_records_to_parquet(&parquet_file, &records).expect("Failed to write discovery records");
+
+    let creature = CreatureJson {
+        input: 1,
+        output: 1,
+        neurons: vec![
+            NeuronJson {
+                uuid: "input-0".to_string(),
+                neuron_type: "input".to_string(),
+                squash: "IDENTITY".to_string(),
+                bias: 0.0,
+            },
+            NeuronJson {
+                uuid: "output-0".to_string(),
+                neuron_type: "output".to_string(),
+                squash: "IDENTITY".to_string(),
+                bias: 0.0,
+            },
+        ],
+        synapses: vec![SynapseJson {
+            from_uuid: "input-0".to_string(),
+            to_uuid: "output-0".to_string(),
+            weight: 0.4,
+            synapse_type: None,
+        }],
+    };
+
+    let input = AnalyzeSynapsesInput {
+        parquet_file,
+        creature,
+        focus_neurons: vec!["output-0".to_string(), "output-0".to_string()],
+        max_candidates: None,
+        analysis_deadline_ms: None,
+        random_seed: None,
+    };
+
+    let err = analyze_synapses(&input)
+        .expect_err("Synapse analysis should refuse duplicate focus neurons");
+    let message = format!("{err}");
+    assert!(
+        message.contains("duplicate focus neurons"),
+        "Expected duplicate focus error, got: {message}",
+    );
+}
+
+#[test]
+fn analyze_synapses_reports_eligible_sources_correctly_for_non_input_neurons() {
+    skip_if_no_gpu!();
+    // Test that non-input neurons with valid creature structure always report
+    // eligible sources correctly, not "no eligible sources" when sources exist
+    let temp_dir = tempdir().expect("Failed to create temporary directory");
+    let parquet_path = temp_dir.path().join("records.parquet");
+    let parquet_file = parquet_path
+        .to_str()
+        .expect("Temporary path should be valid UTF-8")
+        .to_string();
+
+    let sample_count = 100;
+    let mut records = Vec::new();
+    for obs_index in 0..sample_count {
+        // Input neuron records (observations)
+        records.push(DiscoverRecord::new(
+            obs_index,
+            "input-0".to_string(),
+            Some(0.0),
+            0.25,
+            vec![0.1],
+        ));
+        records.push(DiscoverRecord::new(
+            obs_index,
+            "input-1".to_string(),
+            Some(0.0),
+            0.3,
+            vec![0.2],
+        ));
+        // Hidden neuron records
+        records.push(DiscoverRecord::new(
+            obs_index,
+            "hidden-0".to_string(),
+            Some(0.0),
+            0.5,
+            vec![0.15],
+        ));
+        // Output neuron records
+        records.push(DiscoverRecord::new(
+            obs_index,
+            "output-0".to_string(),
+            Some(0.0),
+            0.6,
+            vec![0.05],
+        ));
+    }
+
+    write_records_to_parquet(&parquet_file, &records).expect("Failed to write discovery records");
+
+    let creature = CreatureJson {
+        input: 2,
+        output: 1,
+        neurons: vec![
+            NeuronJson {
+                uuid: "hidden-0".to_string(),
+                neuron_type: "hidden".to_string(),
+                squash: "IDENTITY".to_string(),
+                bias: 0.0,
+            },
+            NeuronJson {
+                uuid: "constant-0".to_string(),
+                neuron_type: "constant".to_string(),
+                squash: "IDENTITY".to_string(),
+                bias: 1.0,
+            },
+            NeuronJson {
+                uuid: "output-0".to_string(),
+                neuron_type: "output".to_string(),
+                squash: "IDENTITY".to_string(),
+                bias: 0.0,
+            },
+        ],
+        synapses: vec![
+            // Hidden neuron already connected to input-0
+            SynapseJson {
+                from_uuid: "input-0".to_string(),
+                to_uuid: "hidden-0".to_string(),
+                weight: 0.4,
+                synapse_type: None,
+            },
+            // Output neuron already connected to hidden-0
+            SynapseJson {
+                from_uuid: "hidden-0".to_string(),
+                to_uuid: "output-0".to_string(),
+                weight: 0.5,
+                synapse_type: None,
+            },
+        ],
+    };
+
+    let input = AnalyzeSynapsesInput {
+        parquet_file,
+        creature,
+        focus_neurons: vec!["hidden-0".to_string(), "output-0".to_string()],
+        max_candidates: None,
+        analysis_deadline_ms: None,
+        random_seed: None,
+    };
+
+    let result = analyze_synapses(&input).expect("Synapse analysis should succeed");
+
+    // Check diagnostics for hidden-0
+    // hidden-0 should have eligible sources (input-1 is not connected yet)
+    // So it should NOT report "no eligible sources"
+    let hidden_diag = result
+        .no_candidate_reasons
+        .iter()
+        .find(|summary| summary.target_uuid == "hidden-0");
+
+    if let Some(diag) = hidden_diag {
+        assert!(
+            diag.reason != SynapseNoCandidateReason::NoEligibleSources,
+            "hidden-0 should have eligible sources (input-1 is available), but got: {:?}",
+            diag.reason
+        );
+        assert!(
+            diag.evaluated_candidates > 0,
+            "hidden-0 should have evaluated at least one candidate (input-1), but evaluated_candidates is {}",
+            diag.evaluated_candidates
+        );
+    }
+
+    // Check diagnostics for output-0
+    // output-0 should have eligible sources (input-0, input-1 are available)
+    // So it should NOT report "no eligible sources"
+    let output_diag = result
+        .no_candidate_reasons
+        .iter()
+        .find(|summary| summary.target_uuid == "output-0");
+
+    if let Some(diag) = output_diag {
+        assert!(
+            diag.reason != SynapseNoCandidateReason::NoEligibleSources,
+            "output-0 should have eligible sources (input-0, input-1 are available), but got: {:?}",
+            diag.reason
+        );
+        assert!(
+            diag.evaluated_candidates > 0,
+            "output-0 should have evaluated at least one candidate, but evaluated_candidates is {}",
+            diag.evaluated_candidates
+        );
+    }
+}
+
+#[test]
+fn analyze_synapses_reports_fully_connected_neuron_explicitly() {
+    skip_if_no_gpu!();
+    // Test that a neuron connected to ALL eligible sources is explicitly reported
+    let temp_dir = tempdir().expect("Failed to create temporary directory");
+    let parquet_path = temp_dir.path().join("records.parquet");
+    let parquet_file = parquet_path
+        .to_str()
+        .expect("Temporary path should be valid UTF-8")
+        .to_string();
+
+    let sample_count = 100;
+    let mut records = Vec::new();
+    for obs_index in 0..sample_count {
+        // Input neuron records (observations)
+        records.push(DiscoverRecord::new(
+            obs_index,
+            "input-0".to_string(),
+            Some(0.0),
+            0.25,
+            vec![0.1],
+        ));
+        records.push(DiscoverRecord::new(
+            obs_index,
+            "input-1".to_string(),
+            Some(0.0),
+            0.3,
+            vec![0.2],
+        ));
+        // Hidden neuron records
+        records.push(DiscoverRecord::new(
+            obs_index,
+            "hidden-0".to_string(),
+            Some(0.0),
+            0.5,
+            vec![0.15],
+        ));
+        // Output neuron records
+        records.push(DiscoverRecord::new(
+            obs_index,
+            "output-0".to_string(),
+            Some(0.0),
+            0.6,
+            vec![0.05],
+        ));
+    }
+
+    write_records_to_parquet(&parquet_file, &records).expect("Failed to write discovery records");
+
+    // Create a creature where hidden-0 is connected to ALL eligible sources
+    // (both input-0 and input-1)
+    let creature = CreatureJson {
+        input: 2,
+        output: 1,
+        neurons: vec![
+            NeuronJson {
+                uuid: "hidden-0".to_string(),
+                neuron_type: "hidden".to_string(),
+                squash: "IDENTITY".to_string(),
+                bias: 0.0,
+            },
+            NeuronJson {
+                uuid: "output-0".to_string(),
+                neuron_type: "output".to_string(),
+                squash: "IDENTITY".to_string(),
+                bias: 0.0,
+            },
+        ],
+        synapses: vec![
+            // hidden-0 is connected to ALL eligible sources (input-0 and input-1)
+            SynapseJson {
+                from_uuid: "input-0".to_string(),
+                to_uuid: "hidden-0".to_string(),
+                weight: 0.4,
+                synapse_type: None,
+            },
+            SynapseJson {
+                from_uuid: "input-1".to_string(),
+                to_uuid: "hidden-0".to_string(),
+                weight: 0.5,
+                synapse_type: None,
+            },
+        ],
+    };
+
+    let input = AnalyzeSynapsesInput {
+        parquet_file,
+        creature,
+        focus_neurons: vec!["hidden-0".to_string()],
+        max_candidates: None,
+        analysis_deadline_ms: None,
+        random_seed: None,
+    };
+
+    let result = analyze_synapses(&input).expect("Synapse analysis should succeed");
+
+    // hidden-0 should be reported as having no eligible sources
+    // because it's connected to ALL eligible sources (both inputs)
+    let hidden_diag = result
+        .no_candidate_reasons
+        .iter()
+        .find(|summary| summary.target_uuid == "hidden-0");
+
+    assert!(
+        hidden_diag.is_some(),
+        "hidden-0 should have diagnostics since it's fully connected"
+    );
+
+    if let Some(diag) = hidden_diag {
+        assert_eq!(
+            diag.reason,
+            SynapseNoCandidateReason::NoEligibleSources,
+            "hidden-0 should report NoEligibleSources since it's connected to all eligible sources"
+        );
+        assert_eq!(
+            diag.evaluated_candidates, 0,
+            "hidden-0 should have 0 evaluated candidates since all sources are already connected"
+        );
+    }
+}
+
+#[test]
+fn analyze_synapses_requires_focus_targets() {
+    let creature = CreatureJson {
+        input: 0,
+        output: 1,
+        neurons: vec![NeuronJson {
+            uuid: "output-0".to_string(),
+            neuron_type: "output".to_string(),
+            squash: "IDENTITY".to_string(),
+            bias: 0.0,
+        }],
+        synapses: Vec::new(),
+    };
+
+    let input = AnalyzeSynapsesInput {
+        parquet_file: "unused.parquet".to_string(),
+        creature,
+        focus_neurons: Vec::new(),
+        max_candidates: None,
+        analysis_deadline_ms: None,
+        random_seed: None,
+    };
+
+    let err =
+        analyze_synapses(&input).expect_err("Synapse analysis should refuse empty focus lists");
+    let message = format!("{err}");
+    assert!(
+        message.contains("at least one focus neuron"),
+        "Expected missing focus error, got: {message}",
+    );
+}
+
+#[test]
+fn analyze_synapses_reports_diagnostics_when_no_candidates() {
+    skip_if_no_gpu!();
+    let temp_dir = tempdir().expect("Failed to create temporary directory");
+    let parquet_path = temp_dir.path().join("records.parquet");
+    let parquet_file = parquet_path
+        .to_str()
+        .expect("Temporary path should be valid UTF-8")
+        .to_string();
+
+    let mut records = Vec::new();
+    for obs_index in 0..16 {
+        records.push(DiscoverRecord::new(
+            obs_index,
+            "output-0".to_string(),
+            Some(0.0),
+            0.25,
+            vec![0.05],
+        ));
+    }
+    write_records_to_parquet(&parquet_file, &records).expect("Failed to write discovery records");
+
+    let creature = CreatureJson {
+        input: 0,
+        output: 1,
+        neurons: vec![NeuronJson {
+            uuid: "output-0".to_string(),
+            neuron_type: "output".to_string(),
+            squash: "IDENTITY".to_string(),
+            bias: 0.0,
+        }],
+        synapses: Vec::new(),
+    };
+
+    let input = AnalyzeSynapsesInput {
+        parquet_file: parquet_file.clone(),
+        creature,
+        focus_neurons: vec!["output-0".to_string()],
+        max_candidates: None,
+        analysis_deadline_ms: None,
+        random_seed: None,
+    };
+
+    let result =
+        analyze_synapses(&input).expect("Synapse analysis should succeed even without candidates");
+    assert!(
+        result.helpful_synapses.is_empty(),
+        "Expected no helpful candidates when there are no eligible sources"
+    );
+    let reason = result
+        .no_candidate_reasons
+        .first()
+        .map(|summary| summary.reason.clone());
+    assert!(
+        matches!(reason, Some(SynapseNoCandidateReason::NoEligibleSources)),
+        "Expected diagnostics to explain missing candidates"
+    );
+}
+
+#[test]
+fn analyze_synapses_stops_harmful_processing_after_deadline() {
+    skip_if_no_gpu!();
+    use rayon::ThreadPoolBuilder;
+    use std::sync::Arc;
+    // Use a private Rayon pool so the deadline override is isolated from other parallel tests.
+    let pool = Arc::new(
+        ThreadPoolBuilder::new()
+            .num_threads(4)
+            .build()
+            .expect("Failed to build Rayon pool"),
+    );
+    let _deadline_guard = deadline_override::DeadlineOverrideGuard::with_sequence_for_pool(
+        vec![
+            false, false, false, false, false, false, true, false, false, false,
+        ],
+        Some(Arc::clone(&pool)),
+    );
+
+    let temp_dir = tempdir().expect("Failed to create temporary directory");
+    let parquet_path = temp_dir.path().join("records.parquet");
+    let parquet_file = parquet_path
+        .to_str()
+        .expect("Temporary path should be valid UTF-8")
+        .to_string();
+
+    let mut records = Vec::new();
+    for obs_index in 0..16 {
+        records.push(DiscoverRecord::new(
+            obs_index,
+            "input-0".to_string(),
+            Some(0.0),
+            1.0,
+            vec![0.0],
+        ));
+        records.push(DiscoverRecord::new(
+            obs_index,
+            "output-0".to_string(),
+            Some(0.0),
+            0.5,
+            vec![0.2],
+        ));
+        records.push(DiscoverRecord::new(
+            obs_index,
+            "output-1".to_string(),
+            Some(0.0),
+            0.5,
+            vec![0.2],
+        ));
+    }
+    write_records_to_parquet(&parquet_file, &records).expect("Failed to write discovery records");
+
+    let creature = CreatureJson {
+        input: 1,
+        output: 2,
+        neurons: vec![
+            NeuronJson {
+                uuid: "output-0".to_string(),
+                neuron_type: "output".to_string(),
+                squash: "IDENTITY".to_string(),
+                bias: 0.0,
+            },
+            NeuronJson {
+                uuid: "output-1".to_string(),
+                neuron_type: "output".to_string(),
+                squash: "IDENTITY".to_string(),
+                bias: 0.0,
+            },
+        ],
+        synapses: vec![
+            SynapseJson {
+                from_uuid: "input-0".to_string(),
+                to_uuid: "output-0".to_string(),
+                weight: 0.8,
+                synapse_type: None,
+            },
+            SynapseJson {
+                from_uuid: "input-0".to_string(),
+                to_uuid: "output-1".to_string(),
+                weight: 0.6,
+                synapse_type: None,
+            },
+        ],
+    };
+
+    let input = AnalyzeSynapsesInput {
+        parquet_file,
+        creature,
+        focus_neurons: vec!["output-0".to_string(), "output-1".to_string()],
+        max_candidates: None,
+        analysis_deadline_ms: None,
+        random_seed: None,
+    };
+
+    let result = pool
+        .install(|| analyze_synapses(&input))
+        .expect("Synapse analysis should complete even when the deadline triggers");
+
+    // With parallel processing, deadline detection order is non-deterministic because
+    // multiple threads call deadline_passed() concurrently. The timeout mechanism is
+    // approximate - once any thread detects the deadline, analysis_timed_out is set
+    // and processing should stop. However, some threads may have already started
+    // processing harmful synapses before the deadline was detected.
+    //
+    // The key requirement is that the analysis completes successfully and respects
+    // the deadline approximately. Since timeout is approximate, we verify that:
+    // 1. The analysis completes without panicking
+    // 2. The result structure is valid
+    // 3. We don't process more harmful synapses than exist (sanity check)
+    //
+    // In this test setup, we have 2 focus neurons, each with 1 harmful synapse (2 total).
+    // The deadline sequence [false x6, true, ...] should cause early termination,
+    // but with parallel processing, the exact point of termination is non-deterministic.
+    let max_possible_harmful = 2; // 2 focus neurons × 1 harmful synapse each
+    assert!(
+        result.harmful_synapses.len() <= max_possible_harmful,
+        "Should not process more harmful synapses than exist. \
+         Got {} harmful synapses, max possible is {}",
+        result.harmful_synapses.len(),
+        max_possible_harmful
+    );
+    // The deadline mechanism is approximate, so we accept any result as long as
+    // the analysis completes and doesn't exceed reasonable bounds
+}
+
+#[test]
+fn analyze_all_runs_synapse_and_neuron_phases() {
+    skip_if_no_gpu!();
+    let temp_dir = tempdir().expect("Failed to create temporary directory");
+    let parquet_path = temp_dir.path().join("records.parquet");
+    let parquet_file = parquet_path
+        .to_str()
+        .expect("Temporary path should be valid UTF-8")
+        .to_string();
+
+    let mut records = Vec::new();
+    for obs_index in 0..(MIN_NEURON_SAMPLE_COUNT as u32 + 1) {
+        records.push(DiscoverRecord::new(
+            obs_index,
+            "input-0".to_string(),
+            Some(0.0),
+            1.0,
+            vec![0.0],
+        ));
+        records.push(DiscoverRecord::new(
+            obs_index,
+            "output-0".to_string(),
+            Some(0.0),
+            0.5,
+            vec![0.2],
+        ));
+    }
+    write_records_to_parquet(&parquet_file, &records).expect("Failed to write discovery records");
+
+    let creature = CreatureJson {
+        input: 1,
+        output: 1,
+        neurons: vec![NeuronJson {
+            uuid: "output-0".to_string(),
+            neuron_type: "output".to_string(),
+            squash: "IDENTITY".to_string(),
+            bias: 0.0,
+        }],
+        synapses: vec![SynapseJson {
+            from_uuid: "input-0".to_string(),
+            to_uuid: "output-0".to_string(),
+            weight: 0.4,
+            synapse_type: None,
+        }],
+    };
+
+    let input = AnalyzeAllInput {
+        parquet_file,
+        creature,
+        focus_neurons: vec!["output-0".to_string()],
+        max_synapse_candidates: Some(5),
+        max_neuron_candidates: Some(5),
+        analysis_deadline_ms: None,
+        include_synapse_analysis: Some(true),
+        include_neuron_analysis: Some(true),
+        random_seed: None,
+    };
+
+    let result = analyze_all(&input).expect("Combined analysis should succeed");
+    assert!(result.synapse.is_some(), "Synapse phase should run");
+    assert!(result.neuron.is_some(), "Neuron phase should run");
+}
+
+#[test]
+fn analyze_neurons_reports_diagnostics_when_no_candidates() {
+    skip_if_no_gpu!();
+    let temp_dir = tempdir().expect("Failed to create temporary directory");
+    let parquet_path = temp_dir.path().join("records.parquet");
+    let parquet_file = parquet_path
+        .to_str()
+        .expect("Temporary path should be valid UTF-8")
+        .to_string();
+
+    let mut records = Vec::new();
+    for obs_index in 0..(MIN_NEURON_SAMPLE_COUNT as u32 + 1) {
+        records.push(DiscoverRecord::new(
+            obs_index,
+            "output-0".to_string(),
+            Some(0.0),
+            0.5,
+            vec![0.2],
+        ));
+    }
+    write_records_to_parquet(&parquet_file, &records).expect("Failed to write discovery records");
+
+    let creature = CreatureJson {
+        input: 0,
+        output: 1,
+        neurons: vec![
+            NeuronJson {
+                uuid: "hidden-source".to_string(),
+                neuron_type: "hidden".to_string(),
+                squash: "IDENTITY".to_string(),
+                bias: 0.0,
+            },
+            NeuronJson {
+                uuid: "output-0".to_string(),
+                neuron_type: "output".to_string(),
+                squash: "IDENTITY".to_string(),
+                bias: 0.0,
+            },
+        ],
+        synapses: Vec::new(),
+    };
+
+    let input = AnalyzeNeuronsInput {
+        parquet_file,
+        creature,
+        focus_neurons: vec!["output-0".to_string()],
+        max_candidates: None,
+        analysis_deadline_ms: None,
+        random_seed: None,
+    };
+
+    let result =
+        analyze_neurons(&input).expect("Neuron analysis should succeed even without candidates");
+    assert!(
+        result.helpful_neurons.is_empty(),
+        "Expected no neuron candidates when the source neuron lacks samples"
+    );
+    let reason = result
+        .no_candidate_reasons
+        .first()
+        .map(|summary| summary.reason.clone());
+    assert!(
+        matches!(reason, Some(NeuronNoCandidateReason::NoSamples)),
+        "Expected diagnostics to explain missing neuron candidates"
+    );
+}
+
+#[test]
+fn analyze_neurons_uses_vertical_timeout_with_randomized_order() {
+    skip_if_no_gpu!();
+    use rayon::ThreadPoolBuilder;
+    use std::sync::Arc;
+
+    // Simulate a deadline that allows at least one focus neuron to start, but
+    // triggers before all are processed. The override sequence is consumed
+    // by calls to `deadline_passed` in order. With randomization, we need
+    // enough false values to allow at least one neuron to start processing.
+    // We provide multiple false values to account for any initialization checks,
+    // then true to stop further processing.
+    // Provide enough "not timed out" checks to allow at least one focus neuron
+    // to begin evaluating sources before we trigger the timeout.
+    let mut deadline_sequence = vec![false; 64];
+    deadline_sequence.push(true);
+    // Use a private pool so the override cannot be consumed by other tests.
+    let pool = Arc::new(
+        ThreadPoolBuilder::new()
+            .num_threads(1)
+            .build()
+            .expect("Failed to build single-threaded Rayon pool"),
+    );
+    let _deadline_guard = deadline_override::DeadlineOverrideGuard::with_sequence_for_pool(
+        deadline_sequence,
+        Some(Arc::clone(&pool)),
+    );
+
+    let temp_dir = tempdir().expect("Failed to create temporary directory");
+    let parquet_path = temp_dir.path().join("records.parquet");
+    let parquet_file = parquet_path
+        .to_str()
+        .expect("Temporary path should be valid UTF-8")
+        .to_string();
+
+    // Provide discovery records for two output neurons but none for the
+    // hidden source. This guarantees that each focus neuron has at least
+    // one eligible upstream source, and that diagnostics can attribute a
+    // `NoSamples` reason once analysis runs.
+    let mut records = Vec::new();
+    for obs_index in 0..(MIN_NEURON_SAMPLE_COUNT as u32 + 1) {
+        records.push(DiscoverRecord::new(
+            obs_index,
+            "output-0".to_string(),
+            Some(0.0),
+            0.5,
+            vec![0.2],
+        ));
+        records.push(DiscoverRecord::new(
+            obs_index,
+            "output-1".to_string(),
+            Some(0.0),
+            0.5,
+            vec![0.2],
+        ));
+    }
+    write_records_to_parquet(&parquet_file, &records).expect("Failed to write discovery records");
+
+    let creature = CreatureJson {
+        input: 0,
+        output: 2,
+        neurons: vec![
+            NeuronJson {
+                uuid: "hidden-source".to_string(),
+                neuron_type: "hidden".to_string(),
+                squash: "IDENTITY".to_string(),
+                bias: 0.0,
+            },
+            NeuronJson {
+                uuid: "output-0".to_string(),
+                neuron_type: "output".to_string(),
+                squash: "IDENTITY".to_string(),
+                bias: 0.0,
+            },
+            NeuronJson {
+                uuid: "output-1".to_string(),
+                neuron_type: "output".to_string(),
+                squash: "IDENTITY".to_string(),
+                bias: 0.0,
+            },
+        ],
+        synapses: Vec::new(),
+    };
+
+    let input = AnalyzeNeuronsInput {
+        parquet_file,
+        creature,
+        focus_neurons: vec!["output-0".to_string(), "output-1".to_string()],
+        max_candidates: None,
+        // Any non-None deadline value will exercise the override sequence.
+        analysis_deadline_ms: Some(1_000_000),
+        random_seed: None,
+    };
+
+    let result = pool
+        .install(|| analyze_neurons(&input))
+        .expect("Neuron analysis should succeed even when the deadline triggers");
+
+    // At least one focus neuron should have evaluated at least one upstream
+    // source before the deadline (vertical timeout behaviour). Since focus
+    // neurons are randomized, we check that at least one of the two neurons
+    // was processed.
+    let processed_neurons: Vec<_> = result
+        .no_candidate_reasons
+        .iter()
+        .filter(|summary| summary.evaluated_sources > 0)
+        .collect();
+
+    // At least one focus neuron should have progressed far enough to attempt
+    // source evaluation, OR we should have produced at least one candidate.
+    let did_any_work = !processed_neurons.is_empty() || !result.helpful_neurons.is_empty();
+    assert!(
+        did_any_work,
+        "At least one focus neuron should do some work before timeout (vertical timeout behaviour)"
+    );
+
+    // Verify that the processed neuron(s) are not reported as having no eligible sources
+    for summary in &processed_neurons {
+        assert!(
+            summary.reason != NeuronNoCandidateReason::NoEligibleSources,
+            "Processed focus neuron should not be reported as having no eligible sources when a timeout occurs"
+        );
+    }
+}
+
+#[test]
+fn analyze_synapses_uses_vertical_timeout_with_randomized_order() {
+    skip_if_no_gpu!();
+    use rayon::ThreadPoolBuilder;
+    use std::sync::Arc;
+
+    // Simulate a deadline that allows at least one focus neuron to start, but
+    // triggers before all are processed. The override sequence is consumed
+    // by calls to `deadline_passed` in order. With randomization, we need
+    // enough false values to allow at least one neuron to start processing.
+    // We provide multiple false values to account for any initialization checks,
+    // then true to stop further processing.
+    // Provide enough "not timed out" checks to allow at least one focus neuron
+    // to begin evaluating candidates before we trigger the timeout.
+    //
+    // The analysis code checks the deadline at several stages (start-of-focus,
+    // source pre-filtering, sample building, batch evaluation). If we trigger
+    // the timeout too early, the vertical-timeout behaviour isn't exercised.
+    let mut deadline_sequence = vec![false; 64];
+    deadline_sequence.push(true);
+    // Use a private pool so the override cannot be consumed by other tests.
+    let pool = Arc::new(
+        ThreadPoolBuilder::new()
+            .num_threads(1)
+            .build()
+            .expect("Failed to build single-threaded Rayon pool"),
+    );
+    let _deadline_guard = deadline_override::DeadlineOverrideGuard::with_sequence_for_pool(
+        deadline_sequence,
+        Some(Arc::clone(&pool)),
+    );
+
+    let temp_dir = tempdir().expect("Failed to create temporary directory");
+    let parquet_path = temp_dir.path().join("records.parquet");
+    let parquet_file = parquet_path
+        .to_str()
+        .expect("Temporary path should be valid UTF-8")
+        .to_string();
+
+    // Provide discovery records for an input neuron and two output neurons.
+    // Records use disjoint obs_index ranges so that each potential synapse
+    // has discovery data but no aligned samples, guaranteeing that the
+    // diagnostics machinery records a `NoSamples` style rejection rather
+    // than treating the target as having no eligible sources.
+    let mut records = Vec::new();
+    for obs_index in 0..16u32 {
+        records.push(DiscoverRecord::new(
+            obs_index,
+            "input-0".to_string(),
+            Some(0.0),
+            1.0,
+            vec![0.0],
+        ));
+    }
+    for obs_index in 100..116u32 {
+        records.push(DiscoverRecord::new(
+            obs_index,
+            "output-0".to_string(),
+            Some(0.0),
+            0.5,
+            vec![0.2],
+        ));
+        records.push(DiscoverRecord::new(
+            obs_index,
+            "output-1".to_string(),
+            Some(0.0),
+            0.5,
+            vec![0.2],
+        ));
+    }
+    write_records_to_parquet(&parquet_file, &records).expect("Failed to write discovery records");
+
+    let creature = CreatureJson {
+        input: 1,
+        output: 2,
+        neurons: vec![
+            NeuronJson {
+                uuid: "output-0".to_string(),
+                neuron_type: "output".to_string(),
+                squash: "IDENTITY".to_string(),
+                bias: 0.0,
+            },
+            NeuronJson {
+                uuid: "output-1".to_string(),
+                neuron_type: "output".to_string(),
+                squash: "IDENTITY".to_string(),
+                bias: 0.0,
+            },
+        ],
+        synapses: Vec::new(),
+    };
+
+    let input = AnalyzeSynapsesInput {
+        parquet_file,
+        creature,
+        focus_neurons: vec!["output-0".to_string(), "output-1".to_string()],
+        max_candidates: None,
+        // Any non-None deadline value will exercise the override sequence.
+        analysis_deadline_ms: None,
+        random_seed: None,
+    };
+
+    let result = pool
+        .install(|| analyze_synapses(&input))
+        .expect("Synapse analysis should succeed even when the deadline triggers");
+
+    // At least one focus neuron should have evaluated at least one upstream
+    // source before the deadline (vertical timeout behaviour). Since focus
+    // neurons are randomized, we check that at least one of the two neurons
+    // was processed.
+    let processed_neurons: Vec<_> = result
+        .no_candidate_reasons
+        .iter()
+        .filter(|summary| summary.evaluated_candidates > 0)
+        .collect();
+
+    // At least one focus neuron should have progressed far enough to attempt
+    // candidate evaluation, OR we should have produced at least one candidate.
+    let did_any_work = !processed_neurons.is_empty()
+        || !result.helpful_synapses.is_empty()
+        || !result.harmful_synapses.is_empty();
+    assert!(
+        did_any_work,
+        "At least one focus neuron should do some work before timeout (vertical timeout behaviour)"
+    );
+
+    // If we observed a processed neuron via diagnostics, it should not be reported
+    // as having no eligible sources.
+    for summary in &processed_neurons {
+        assert!(
+            summary.reason != SynapseNoCandidateReason::NoEligibleSources,
+            "Processed focus neuron should not be reported as having no eligible sources when a timeout occurs"
+        );
+    }
+}
+
+/// Test that positive improvements below threshold are accepted as candidates
+/// This verifies the fix where all positive improvements are candidates, not just those above threshold
+#[test]
+fn analyze_synapses_accepts_positive_improvements_below_threshold() {
+    skip_if_no_gpu!();
+    let temp_dir = tempdir().expect("Failed to create temporary directory");
+    let parquet_path = temp_dir.path().join("records.parquet");
+    let parquet_file = parquet_path
+        .to_str()
+        .expect("Temporary path should be valid UTF-8")
+        .to_string();
+
+    // Create test data that will produce a positive but below-threshold improvement
+    // We need: expected_improvement = (2*w*E[a*e] - w^2*E[a^2]) / E[e^2]
+    // To get ~0.05 improvement with threshold 0.1, we'll use:
+    // - source activation: 0.5 consistently
+    // - target error: 0.1 consistently
+    // - This should produce a positive improvement when weight is chosen appropriately
+    let sample_count = 100;
+    let mut records = Vec::new();
+    for obs_index in 0..sample_count {
+        // Source neuron (input-0) with consistent activation
+        records.push(DiscoverRecord::new(
+            obs_index,
+            "input-0".to_string(),
+            Some(0.0),
+            0.5,    // Consistent activation
+            vec![], // Input neurons don't have errors
+        ));
+        // Target neuron (output-0) with consistent error
+        records.push(DiscoverRecord::new(
+            obs_index,
+            "output-0".to_string(),
+            Some(0.0),
+            0.3,
+            vec![0.1], // Consistent error
+        ));
+    }
+
+    write_records_to_parquet(&parquet_file, &records).expect("Failed to write discovery records");
+
+    let creature = CreatureJson {
+        input: 1,
+        output: 1,
+        neurons: vec![NeuronJson {
+            uuid: "output-0".to_string(),
+            neuron_type: "output".to_string(),
+            squash: "IDENTITY".to_string(),
+            bias: 0.0,
+        }],
+        synapses: Vec::new(), // No existing synapse from input-0 to output-0
+    };
+
+    let input = AnalyzeSynapsesInput {
+        parquet_file,
+        creature,
+        focus_neurons: vec!["output-0".to_string()],
+        max_candidates: None,
+        analysis_deadline_ms: None,
+        random_seed: None,
+    };
+
+    let result = analyze_synapses(&input).expect("Synapse analysis should succeed");
+
+    // The key assertion: positive improvements below threshold should be accepted
+    // We should have at least one helpful synapse candidate (even if improvement < 0.1)
+    // OR if no candidate, it should NOT be due to BelowThreshold for a positive improvement
+    if result.helpful_synapses.is_empty() {
+        // If no candidates, check diagnostics - it should NOT be BelowThreshold for positive improvements
+        let no_candidate = result
+            .no_candidate_reasons
+            .iter()
+            .find(|summary| summary.target_uuid == "output-0");
+
+        if let Some(summary) = no_candidate {
+            // If there's a detail, check that it's not a positive improvement below threshold
+            if let Some(detail) = &summary.detail {
+                if let Some(improvement) = detail.expected_improvement {
+                    if improvement > 0.0 && improvement <= 0.1 {
+                        panic!(
+                            "Positive improvement {:.4} below threshold 0.1 should be accepted as candidate, but was rejected with reason: {:?}",
+                            improvement, summary.reason
+                        );
+                    }
+                }
+            }
+        }
+    } else {
+        // We have candidates - verify at least one has positive improvement
+        // Issue #128: Use expected_creature_score_gain (creature-level, not neuron-level)
+        let has_positive_improvement = result
+            .helpful_synapses
+            .iter()
+            .any(|synapse| synapse.expected_creature_score_gain > 0.0);
+
+        assert!(
+            has_positive_improvement,
+            "Should have at least one candidate with positive improvement"
+        );
+    }
+}
+
+/// Test that non-positive improvements (<= 0.0) are still rejected
+#[test]
+fn analyze_synapses_rejects_non_positive_improvements() {
+    skip_if_no_gpu!();
+    let temp_dir = tempdir().expect("Failed to create temporary directory");
+    let parquet_path = temp_dir.path().join("records.parquet");
+    let parquet_file = parquet_path
+        .to_str()
+        .expect("Temporary path should be valid UTF-8")
+        .to_string();
+
+    // Create test data that will produce a non-positive improvement
+    // Use mismatched activations/errors that result in negative or zero improvement
+    let sample_count = 100;
+    let mut records = Vec::new();
+    for obs_index in 0..sample_count {
+        // Source neuron with activation
+        records.push(DiscoverRecord::new(
+            obs_index,
+            "input-0".to_string(),
+            Some(0.0),
+            1.0,
+            vec![],
+        ));
+        // Target neuron with error that doesn't correlate well (will produce negative improvement)
+        records.push(DiscoverRecord::new(
+            obs_index,
+            "output-0".to_string(),
+            Some(0.0),
+            0.0,
+            vec![-0.1], // Negative error when source is positive
+        ));
+    }
+
+    write_records_to_parquet(&parquet_file, &records).expect("Failed to write discovery records");
+
+    let creature = CreatureJson {
+        input: 1,
+        output: 1,
+        neurons: vec![NeuronJson {
+            uuid: "output-0".to_string(),
+            neuron_type: "output".to_string(),
+            squash: "IDENTITY".to_string(),
+            bias: 0.0,
+        }],
+        synapses: Vec::new(),
+    };
+
+    let input = AnalyzeSynapsesInput {
+        parquet_file,
+        creature,
+        focus_neurons: vec!["output-0".to_string()],
+        max_candidates: None,
+        analysis_deadline_ms: None,
+        random_seed: None,
+    };
+
+    let result = analyze_synapses(&input).expect("Synapse analysis should succeed");
+
+    // Non-positive improvements should be rejected (not appear in helpful_synapses)
+    // Even though we now accept positive improvements below threshold, we still reject <= 0.0
+    // Issue #128: Use expected_creature_score_gain (creature-level metric)
+    let has_non_positive = result
+        .helpful_synapses
+        .iter()
+        .any(|synapse| synapse.expected_creature_score_gain <= 0.0);
+
+    assert!(
+        !has_non_positive,
+        "Should not have any candidates with non-positive improvement (<= 0.0)"
+    );
+}
+
+/// Test that positive improvements above threshold are still accepted (regression test)
+#[test]
+fn analyze_synapses_accepts_positive_improvements_above_threshold() {
+    skip_if_no_gpu!();
+    let temp_dir = tempdir().expect("Failed to create temporary directory");
+    let parquet_path = temp_dir.path().join("records.parquet");
+    let parquet_file = parquet_path
+        .to_str()
+        .expect("Temporary path should be valid UTF-8")
+        .to_string();
+
+    // Create test data that will produce a positive improvement above threshold
+    // Use strong correlation between source activation and target error
+    let sample_count = 100;
+    let mut records = Vec::new();
+    for obs_index in 0..sample_count {
+        // Source neuron with strong activation
+        records.push(DiscoverRecord::new(
+            obs_index,
+            "input-0".to_string(),
+            Some(0.0),
+            1.0,
+            vec![],
+        ));
+        // Target neuron with error that correlates positively
+        records.push(DiscoverRecord::new(
+            obs_index,
+            "output-0".to_string(),
+            Some(0.0),
+            0.5,
+            vec![0.2], // Positive error when source is positive
+        ));
+    }
+
+    write_records_to_parquet(&parquet_file, &records).expect("Failed to write discovery records");
+
+    let creature = CreatureJson {
+        input: 1,
+        output: 1,
+        neurons: vec![NeuronJson {
+            uuid: "output-0".to_string(),
+            neuron_type: "output".to_string(),
+            squash: "IDENTITY".to_string(),
+            bias: 0.0,
+        }],
+        synapses: Vec::new(),
+    };
+
+    let input = AnalyzeSynapsesInput {
+        parquet_file,
+        creature,
+        focus_neurons: vec!["output-0".to_string()],
+        max_candidates: None,
+        analysis_deadline_ms: None,
+        random_seed: None,
+    };
+
+    let result = analyze_synapses(&input).expect("Synapse analysis should succeed");
+
+    // Positive improvements above threshold should definitely be accepted
+    // This is a regression test to ensure we didn't break existing behavior
+    // Issue #128: Use expected_creature_score_gain (creature-level metric)
+    let has_above_threshold = result
+        .helpful_synapses
+        .iter()
+        .any(|synapse| synapse.expected_creature_score_gain > 0.1);
+
+    // Note: This test may pass even if no candidates are found due to other reasons
+    // (e.g., no samples, zero improvement). The key is that if we have candidates,
+    // they should include positive improvements above threshold.
+    if !result.helpful_synapses.is_empty() {
+        assert!(
+            has_above_threshold
+                || result
+                    .helpful_synapses
+                    .iter()
+                    .any(|s| s.expected_creature_score_gain > 0.0),
+            "Should have candidates with positive improvement (above or below threshold)"
+        );
+    }
+}
